# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type Priority = (typeof Priority)[keyof typeof Priority];

export const Priority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: string;
  title: string;
  description?: string | null;
  isCompleted: boolean;
  priority: Priority;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /** @maxLength 2000 */
  description?: string | null;
  priority?: Priority;
  dueDate?: string | null;
}

export interface UpdateTaskInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  /** @maxLength 2000 */
  description?: string | null;
  isCompleted?: boolean;
  priority?: Priority;
  dueDate?: string | null;
}

export interface TaskSummary {
  total: number;
  completed: number;
  pending: number;
  completedToday: number;
  dueToday: number;
  overdue: number;
  completionRate: number;
}

/**
 * Bad request
 */
export type BadRequestResponse = Error;

/**
 * Unauthorized
 */
export type UnauthorizedResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type ListTasksParams = {
  status?: ListTasksStatus;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  all: "all",
  pending: "pending",
  completed: "completed",
} as const;