# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for the signed-in user
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/summary:
    get:
      operationId: getTaskSummary
      tags: [tasks]
      summary: Aggregate stats for the signed-in user
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Priority:
      type: string
      enum: [low, medium, high]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        isCompleted:
          type: boolean
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, isCompleted, priority, createdAt, updatedAt]
    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
          nullable: true
      required: [title]
    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        isCompleted:
          type: boolean
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
          nullable: true
    TaskSummary:
      type: object
      properties:
        total:
          type: integer
        completed:
          type: integer
        pending:
          type: integer
        completedToday:
          type: integer
        dueToday:
          type: integer
        overdue:
          type: integer
        completionRate:
          type: number
      required: [total, completed, pending, completedToday, dueToday, overdue, completionRate]