# lib/api-zod/src/generated/types/priority.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export type Priority = (typeof Priority)[keyof typeof Priority];

export const Priority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;