# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { Priority } from "./priority";

export interface Task {
  id: string;
  title: string;
  description?: string | null;
  isCompleted: boolean;
  priority: Priority;
  dueDate?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}