# lib/api-zod/src/generated/types/taskSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskSummary {
  total: number;
  completed: number;
  pending: number;
  completedToday: number;
  dueToday: number;
  overdue: number;
  completionRate: number;
}