# lib/api-zod/src/generated/types/updateTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { Priority } from "./priority";

export interface UpdateTaskInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  /** @maxLength 2000 */
  description?: string | null;
  isCompleted?: boolean;
  priority?: Priority;
  dueDate?: Date | null;
}